Fix response parsing bug in Ruby 2.0
The signature of Net::IMAP::ResponseParser#msg_att has changed in ruby
2.0.0. You can see the relevant issue here:
http://bugs.ruby-lang.org/issues/5692

and the diff here:
http://bugs.ruby-lang.org/projects/ruby-trunk/repository/revisions/34010

This change caused some tests to fail when using gmail-imap 1.0.1 with
Ruby 2.0. This patch adjusts to the new function signature while still
allowing the library to work in versions of ruby older than 2.0. I've
run the test suite on 1.9.3-p448 and 2.0.0-p247 and verified that
everything works. The fix is pretty gross, but I didn't see a better way
to do it.